clean up config
improve handling
