+++ addons/h1/config/weapons.hpp
@@ -1,0 +1,238 @@
+class CfgMagazines
+{
+		//GAU-21
+	class 100Rnd_127x99_mag_Tracer_Red;
+	class 600Rnd_127x99_mag_Tracer_Red: 100Rnd_127x99_mag_Tracer_Red
+	{
+		count=600;
+		tracersEvery=5;
+	};
+	//GAU-17
+	class 2000Rnd_65x39_Belt_Tracer_Red;
+	class 4000Rnd_762x51_Belt_Tracer_Red: 2000Rnd_65x39_Belt_Tracer_Red
+	{
+		count=4000;
+		tracersEvery=5;
+	};
+};
+class CfgWeapons
+{
+	// M3M GAU-21
+	class HMG_01;
+	class HMG_M2: HMG_01
+	{
+		class manual;
+	};
+	class M3M_GAU21: HMG_M2
+	{
+		displayName="GAU-21";
+		magazines[]=
+		{
+			"600Rnd_127x99_mag_Tracer_Red"
+		};
+		class manual: manual
+		{
+			reloadTime=0.054499999;
+		};
+		class GunParticles
+		{
+			class effect1
+			{
+				positionName="GAU21L_End";
+				directionName="GAU21L_Beg";
+				effectName="MachineGunCloud";
+			};
+			class effect2
+			{
+				positionName="GAU21L_eject_pos";
+				directionName="GAU21L_eject_dir";
+				effectName="MachineGunEject";
+			};
+			class effect3
+			{
+				positionName="GAU21L_eject_pos";
+				directionName="GAU21L_eject_dir";
+				effectName="MachineGunCartridge2";
+			};
+		};
+	};
+	// M134 GAU-17
+	class LMG_Minigun;
+	class LMG_Minigun_Transport: LMG_Minigun
+	{
+		class manual;
+		class far;
+		class medium;
+		class close;
+		class short;
+	};
+	class M134_GAU17: LMG_Minigun_Transport
+	{
+		magazines[]=
+		{
+			"4000Rnd_762x51_Belt_Tracer_Red"
+		};
+		class gunParticles
+		{
+			class effect1
+			{
+				positionName="cartridge_pos";
+				directionName="cartridge_dir";
+				effectName="MachineGunCartridge1";
+			};
+			class effect2
+			{
+				positionName="muzzle_1";
+				directionName="chamber_1";
+				effectName="MachineGun1";
+			};
+		};
+		class manual: manual
+		{
+			reloadTime=0.015;
+			dispersion=0.0099999998;
+			sounds[]=
+			{
+				"StandardSound"
+			};
+			class StandardSound
+			{
+				begin1[]=
+				{
+					"h1\sounds\m134\m134",
+					20,
+					0.89999998,
+					1500,
+					{25704,27159}
+				};
+				soundBegin[]=
+				{
+					"begin1",
+					10
+				};
+			};
+			ffMagnitude=0.5;
+			ffFrequency=11;
+			ffCount=6;
+			burst=20;
+			soundContinuous=1;
+			autoFire=1;
+		};
+		class far: far
+		{
+			sounds[]=
+			{
+				"StandardSound"
+			};
+			class StandardSound
+			{
+				begin1[]=
+				{
+					"h1\sounds\m134\m134",
+					20,
+					0.89999998,
+					1500,
+					{25704,27159}
+				};
+				soundBegin[]=
+				{
+					"begin1",
+					10
+				};
+			};
+			ffMagnitude=0.5;
+			ffFrequency=11;
+			ffCount=6;
+			burst=20;
+			soundContinuous=1;
+			autoFire=1;
+		};
+		class medium: medium
+		{
+			sounds[]=
+			{
+				"StandardSound"
+			};
+			class StandardSound
+			{
+				begin1[]=
+				{
+					"h1\sounds\m134\m134",
+					20,
+					0.89999998,
+					1500,
+					{25704,27159}
+				};
+				soundBegin[]=
+				{
+					"begin1",
+					10
+				};
+			};
+			ffMagnitude=0.5;
+			ffFrequency=11;
+			ffCount=6;
+			burst=20;
+			soundContinuous=1;
+			autoFire=1;
+		};
+		class close: close
+		{
+			sounds[]=
+			{
+				"StandardSound"
+			};
+			class StandardSound
+			{
+				begin1[]=
+				{
+					"h1\sounds\m134\m134",
+					20,
+					0.89999998,
+					1500,
+					{25704,27159}
+				};
+				soundBegin[]=
+				{
+					"begin1",
+					10
+				};
+			};
+			ffMagnitude=0.5;
+			ffFrequency=11;
+			ffCount=6;
+			burst=20;
+			soundContinuous=1;
+			autoFire=1;
+		};
+		class short: short
+		{
+			sounds[]=
+			{
+				"StandardSound"
+			};
+			class StandardSound
+			{
+				begin1[]=
+				{
+					"h1\sounds\m134\m134",
+					20,
+					0.89999998,
+					1500,
+					{25704,27159}
+				};
+				soundBegin[]=
+				{
+					"begin1",
+					10
+				};
+			};
+			ffMagnitude=0.5;
+			ffFrequency=11;
+			ffCount=6;
+			burst=20;
+			soundContinuous=1;
+			autoFire=1;
+		};
+	};
+};
